requirements.txt
# Core dependencies
requests>=2.25.0
beautifulsoup4>=4.9.0
spacy>=3.0.0
chardet>=4.0.0

# Document support
PyPDF2>=2.0.0
python-docx>=0.8.11
openpyxl>=3.0.7

# OCR support
pytesseract>=0.3.8
Pillow>=8.0.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.5b2
flake8>=3.9.0
isort>=5.9.0
build>=0.7.0
twine>=3.4.2